Refactor Config and ConfigManager: Change configVersion type to double and update version comparison logic

+++ src/main/java/com/hanhy06/embellish_chat/config/Config.java
@@ -8,7 +8,7 @@ import java.util.*;
 import static java.util.Map.entry;
 
 public record Config(
-        int configVersion,
+        double configVersion,
 
         //style
         HashMap<String,List<StylingRule>> stylingRules,
